// src/app/models/mtr-cliente-direccion-update-dto.ts
export class MtrClienteDireccionUpdateDto {
    id: number;
    codigo: string;
    idDireccionCliente: number;
    rifCliente: string;
    rifDireccion: string;
    direccion: string;
    direccion1: string;
    direccion2: string;
    estado: string;
    municipio: string;
    idTipoNegocio: string;
    sector: string;
    ramo: string;
    puntoReferencia: string;
    usuario: string;
}